+++ src/wpapi.js
@@ -1,6 +1,14 @@
-export let WP = null
+import invariants from './invariants'
+
+let WP = null
+
+// Get the internal reference to the wpapi instance
+export function getWP () {
+  return WP
+}
 
 // Set an internal reference to the wpapi instance
 export function setWP (_WP) {
+  invariants.noWPInstance(WP)
   WP = _WP
 }