+++ test/normalise.js
@@ -30,7 +30,11 @@ function normaliserTestData (contentType) {
 }
 
 function setup () {
-  registerContentType({
+  const mockWP = {
+    registerRoute: jest.fn()
+  }
+
+  registerContentType(mockWP, {
     name: 'book',
     plural: 'books',
     slug: 'books'